+++ NEWS.md
@@ -1,3 +1,16 @@
+# ggtibble 1.0.2
+
+* `ggtibble()` now warns if `outercols` are not used in either the `caption` or
+  the `labs` argument (#13).
+* `ggtibble` and `gglist` objects now work with the ggplot2 `%+%` operator (#16)
+* A new `ggsave()` generic function will now enable simpler saving of `ggtibble`
+  and `gglist` objects (unique filenames are required to save).
+* `aes()` and `data.frame()` objects can be added to `ggtibble` and `gglist`
+  objects (#23).
+* `ggsave()` can accept a character vector of all filenames to use when saving
+  (#25).
+* Update testing to work with ggplot2 version 4.0.0
+
 # ggtibble 1.0.1
 
 * `labs` argument to `ggtibble()` can now include `NULL (#6)